support arm64

+++ SDWebImage/SDImageCache.h
@@ -163,7 +163,7 @@ typedef enum SDImageCacheType SDImageCacheType;
 /**
  * Asynchronously calculate the disk cache's size.
  */
-- (void)calculateSizeWithCompletionBlock:(void (^)(NSUInteger fileCount, unsigned long long totalSize))completionBlock;
+- (void)calculateSizeWithCompletionBlock:(void (^)(NSUInteger fileCount, NSUInteger totalSize))completionBlock;
 
 /**
  * Check if image exists in cache already
